refactor: lowercase umd output name and entrypoint
