+++ src/component/interfaces/KeyboardConfiguration.ts
@@ -6,7 +6,7 @@ import { ComponentConfiguration } from "./ComponentConfiguration";
  * @interface
  * @example
  * ```
- * var viewer = new Mapillary.Viewer({
+ * var viewer = new mapillary.Viewer({
  *     ...
  *     component: {
  *         keyboard: {
