+++ src/component/marker/marker/SimpleMarker.ts
@@ -16,11 +16,11 @@ import { Marker } from "./Marker";
  *
  * @example
  * ```
- * var defaultMarker = new Mapillary.MarkerComponent.SimpleMarker(
+ * var defaultMarker = new mapillary.MarkerComponent.SimpleMarker(
  *     "id-1",
  *     { lat: 0, lon: 0, });
  *
- * var interactiveMarker = new Mapillary.MarkerComponent.SimpleMarker(
+ * var interactiveMarker = new mapillary.MarkerComponent.SimpleMarker(
  *     "id-2",
  *     { lat: 0, lon: 0, },
  *     {
